requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
cycler==0.10.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
kiwisolver==1.1.0
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.4
Pillow==6.0.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
python-telegram-bot==11.1.0
scipy==1.1.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
